requirements.txt
ephem
configparser
Flask
paho-mqtt
pigpio